feat(handler): move handler code to handler pkg
